% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layout.R
\name{get_layout}
\alias{get_layout}
\title{Get a label layout.}
\usage{
get_layout(label.type)
}
\arguments{
\item{label.type}{A character string describing the label vendor, product number and output page size.}
}
\description{
A utility to retrieve pre-formatted label layout configurations.
}
\details{
This function allows easy access to label layout configurations for a range of vendor's label formats. More layouts are being continually added and users are welcome to submit their own. See the package website for more information on the available configurations
}
\examples{
require(idLabelR)

example_spreadsheet = system.file("extdata", "Example.txt", package = "idLabelR")
ids = make_ids(example_spreadsheet)
layout = get_layout("cryobabies_LCRY_1700_A4")
make_pdf(ids, layout)

}
\seealso{
\code{\link[idLabelR]{make_pdf}}.
}
