% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pdf.R
\name{make_pdf}
\alias{make_pdf}
\title{Write a printable PDF file.}
\usage{
make_pdf(ids, layout, byRow = TRUE, printing = NULL, file = "Labels.pdf", ...)
}
\arguments{
\item{ids}{A character vector of IDs as generated by
\code{\link[idLabelR]{make_ids}}.}

\item{layout}{A \code{\link{list}} object describing the layout of labels on
the sheet to be printed. See Details to learn how to customise this.}

\item{byRow}{By default, IDs are repeated for all labels in a row. If
\code{byRow = FALSE}, then IDs are repeated for all labels in a column.}

\item{printing}{An optional character to identify multiple copies of the same
label sheet. Default is to ignore this feature.}

\item{file}{The name of the file to be written to. The default is
\code{"Labels.pdf"} in the current working directory.}

\item{...}{Additional parameters passed to \code{\link[grDevices]{pdf}}.}
}
\description{
A utility to generate a PDF file ready for printing labels.
}
\details{
This function generates a PDF document that can be printed directly to sheets
of labels/stickers for the easy and clear identification of experimental
samples. The default behaviour is to print the same ID across all labels in a
row, with a different ID for each row.
}
\examples{
require(idLabelR)
example_spreadsheet = system.file("extdata", "Example.txt", package = "idLabelR")
ids = make_ids(file = example_spreadsheet)
layout = get_layout("cryobabies_LCRY_1700_A4")
make_pdf(ids, layout)

# Additionally prints a timestamp and identifier
# to distinguish repeated printings of the same page.
make_pdf(ids, layout, printing = "a")

# 17 rows of 5 labels for each ID.
make_pdf(ids, layout, byRow = TRUE, file = "Labels_by_row.pdf")
# 5 columns of 17 labels for each ID.
make_pdf(ids, layout, byRow = FALSE, file = "Labels_by_col.pdf")

}
\seealso{
\code{\link[idLabelR]{make_ids}}. Also see
  \code{\link[grDevices]{pdf}} for detail on additional PDF printing
  parameters.
}
