% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_db.R
\name{read_db}
\alias{read_db}
\title{Read a spreadsheet.}
\usage{
read_db(file, filetype = "auto", ...)
}
\arguments{
\item{file}{The name of the file to be read.}

\item{filetype}{The filetype of the input spreadsheet. Microsoft Excel
(\code{.xlsx} and \code{.xls}) and Open Document Format (\code{.ods}; as
used by OpenOffice/LibreOffice for example) are supported, as are comma-
and tab-delimited formats (\code{.csv}, \code{.tsv}/\code{.txt}). The
default behaviour is to guess the format from the file extension. If this
does not work or you know better than the machine, then this option can be
used to specify the filetype.}

\item{...}{Additional parameters passed to the file reader. The parameters
accepted depend on the input filetype.}
}
\description{
A helper utility to handle reading of spreadsheet files.
}
\details{
This function should not need to be called directly by the user.

The code used to read the file depends on the filetype. The filetype is
determined by the extension given in the \code{file} parameter or can be
overridden by the \code{filetype} parameter. For more information about file
reading parameters, have a look at the documentation for the respective
packages: \code{\link[readxl]{read_excel}} for reading \code{.xlsx} and
\code{.xls} files. \code{\link[readODS]{read_ods}} for reading \code{.ods}
files. This is the Open Document standard used by software such as OpenOffice
and LibreOffice. \code{\link[utils]{read.csv}} for reading \code{.csv} files.
If \code{filetype = 'csv2'}, then \code{\link[utils]{read.csv2}} is used
(where commas are used as decimal separators as is common in Europe).
\code{\link[utils]{read.delim}} for reading \code{.tsv}, \code{.tab},
\code{.txt} files. This is just tab-delimited text.
}
\seealso{
\code{\link[idLabelR]{make_ids}}.
}
