% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_db.R
\name{write_db}
\alias{write_db}
\title{Write a spreadsheet with IDs.}
\usage{
write_db(df, file, filetype, ...)
}
\arguments{
\item{df}{A \code{\link{data.frame}} containing the experiment data.}

\item{file}{The name of the file to be written to.}

\item{filetype}{The filetype of the input spreadsheet. Microsoft Excel
(\code{.xlsx} and \code{.xls}) and Open Document Format (\code{.ods}; as
used by OpenOffice/LibreOffice for example) are supported, as are comma-
and tab-delimited formats (\code{.csv}, \code{.tsv}/\code{.txt}). The
default behaviour is to guess the format from the file extension. If this
does not work or you know better than the machine, then this option can be
used to specify the filetype.}

\item{...}{Additional parameters passed to the file reader. The parameters
accepted depend on the input filetype.}
}
\description{
A helper utility to handle writing of spreadsheet files.
}
\details{
This function should not need to be called directly by the user.

The code used to write the file depends on the filetype. The filetype is
determined by the extension given in the \code{file} parameter or can be
overridden by the \code{filetype} parameter. For more information about file
writing parameters, have a look at the documentation for the respective
packages: \code{\link[openxlsx]{write.xlsx}} for writing \code{.xlsx} files.
Note that an input filetype of \code{.xls} will cause a \code{.xlsx} file to
be written (as \code{.xls} is not supported).
\code{\link[readODS]{write_ods}} for writing \code{.ods} files. This is the
Open Document standard used by software such as OpenOffice and LibreOffice.
\code{\link[utils]{write.csv}} for writing \code{.csv} files. If
\code{filetype = 'csv2'}, then \code{\link[utils]{write.csv2}} is used (where
commas are used as decimal separators as is common in Europe).
\code{\link[utils]{write.table}} for writing \code{.tsv}, \code{.tab},
\code{.txt} files. This is just tab-delimited text. The same file extension
used for the input file will be used to write the output.
}
\seealso{
\code{\link[idLabelR]{make_ids}}.
}
